requirements.txt
requests
numpy
pandas
python-dotenv
tqdm
plotly
lightgbm
scikit-learn
optuna
scipy
pyyaml
openpyxl
